+++ src/elements/Svg.js
@@ -62,6 +62,10 @@ export default class Svg extends Container {
     while (this.node.hasChildNodes()) {
       this.node.removeChild(this.node.lastChild)
     }
+
+    // remove defs reference
+    delete this._defs
+
     return this
   }
 }